requirements.txt
Pillow
djangorestframework
python-decouple
dj-database-url
